Edit comment now works

+++ client/src/Components/CommentComponents/Comment.js
@@ -1,18 +1,59 @@
-import { Button, Card, CardBody, CardFooter, CardHeader } from "reactstrap";
+import { Button, Card, CardBody, CardFooter, CardHeader, Spinner } from "reactstrap";
 import UserPartial from "../UserComponents/UserPartial";
-import { deleteComment } from "../../Modules/commentManager";
+import { deleteComment, editComment } from "../../Modules/commentManager";
+import { useEffect, useState } from "react";
 
 export default function Comment({ comment, board, loggedInUser, retrievePost }) {
     const commentDate = new Date(comment.createdOn).toLocaleDateString()
+
+    const [editState, setEdit] = useState(null);
+    const [editText, setEditText] = useState("");
+
+    useEffect(() => {
+        setEditText(comment.content)
+    }, [comment])
+
+    const editButtonClickHandler = (e) => {
+        e.preventDefault();
+        const editedComment = { ...comment };
+        editedComment.content = editText;
+        if (!editedComment.content.length > 0) {
+            alert("Comment's content must not be empty!")
+            return;
+        }
+        try {
+            editComment(comment.id, editedComment)
+            setEdit(null);
+            retrievePost()
+        }
+        catch (ex) {
+            alert(ex.message)
+        }
+    }
+
+    if (loggedInUser === null || comment === undefined || board === undefined || retrievePost === undefined) {
+        return <Spinner className="app-spinner dark" />
+    }
+
     return <Card>
         <CardBody>
-            <p>{comment.content}</p>
+            {
+                editState === null ?
+                    <h5>{comment.content}</h5>
+                    :
+                    <>
+                        <input type="text" value={editText} onChange={e => { setEditText(e.target.value) }} style={{ width: "100%", height: "100%" }} />
+                        <Button color="info" onClick={e => { editButtonClickHandler(e) }}>Submit Edit</Button>
+                        <Button color="danger" onClick={() => { setEdit(null) }}>Cancel</Button>
+                    </>
+            }
         </CardBody>
         <p>Comment made on: {commentDate}</p>
         <UserPartial userInfo={comment.user} />
         <CardFooter>
+
             {
-                loggedInUser.id === board.boardOwner.id || loggedInUser.id === comment.user.id ?
+                loggedInUser !== null && loggedInUser.id === board.boardOwner.id || loggedInUser.id === comment.user.id ?
                     <Button color="danger" onClick={e => {
                         e.preventDefault();
                         deleteComment(comment.id);
@@ -20,6 +61,18 @@ export default function Comment({ comment, board, loggedInUser, retrievePost }) 
                     }}>Delete</Button>
                     : ""
             }
+
+            {
+                loggedInUser !== null && loggedInUser.id === comment.user.id ?
+                    <Button color="warning" onClick={e => {
+                        e.preventDefault();
+                        setEdit(true);
+                    }}>Edit</Button>
+                    :
+                    ""
+            }
+
+
         </CardFooter>
     </Card>
 }
